requirements.txt
absl-py==2.1.0
anyio==3.7.1
asttokens==2.4.0
astunparse==1.6.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
exceptiongroup==1.2.2
fastapi==0.103.2
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.55.2
gast==0.4.0
google-auth==2.36.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.68.1
h5py==3.11.0
idna==3.10
importlib_metadata==6.7.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
keras==2.13.1
kiwisolver==1.4.7
kt-legacy==1.0.5
libclang==18.1.1
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
nest_asyncio==1.5.7
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
pandas==2.0.3
pillow==10.4.0
plotly==5.24.1
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==1.10.19
pyparsing==3.1.4
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
starlette==0.27.0
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
tenacity==9.0.0
tzdata==2024.2
urllib3==2.2.3
waitress==3.0.0
Werkzeug==3.0.6
wrapt==1.17.0
zipp==3.20.2